Move class QueuePoint to its own file
