+++ .gitignore
@@ -1,2 +1,2 @@
-Queue.o
+*.o
 evalqueue
