+++ QueuePoint.cpp
@@ -1,0 +1,2 @@
+#include "QueuePoint.hpp"
+
